Remap ctrl-t to BrowserHome({ button: 1 })
This is equivalent to clicking the Home button on the toolbar with the
middle mouse button.  This opens a new tab AND loads the home URL, so
the user can set up any home URL via settings, no need to hardcode
Hoplax into the build.

Also, this puts the focus into Hoplax, so it's just perfect for us.